Define a report package with core and per format sub-packages
