+++ report/html/template.go
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package output
+package html
 
-const html = `
+const templateContent = `
 <!doctype html>
 <html lang="en">
 <head>
