+++ report/sonar/types.go
